pyproject.toml
[project]
name = "nesterenkobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "discord>=2.3.2",
    "google-genai>=1.32.0",
    "nextcord>=3.1.1",
    "requests>=2.32.5",
]
